Dion state progress
Partial implementation of DionIdleState
Renamed Dion's "moves" to "actions" to distinguish them from movement related logic

Added DionStageAI which handles selecting actions and holds the data about moves (potentially) unique to each stage
Added DionStageAIController which handles an updated reference to the current DionStageAI for the current stage Dion is on
Commented out the testing stuff in GameManager because it was dumb and bad and just creates errors
DionIdleState's references to other states are handled through direct onready get_node()s instead of indirect and annoying exporting to do the same thing
Added Transition and Idle states to the base DionState script as most or all scripts will use these transitions

Wow that was a lotta notes idk why I actually wrote on everything here. If you see this, know that you are cool and epic and I appreciate you random internet person!